refactor: extract VSegmentControl from pill tabs, update VToggle colors
- Create VSegmentControl as a standalone component matching Figma design
  (contentBackground container, contentInset active bg, 6px segment radius)
- Remove pill/compact variants from VTabs (now underline-only)
- Delete VThemeToggle, replace with VSegmentControl + VTheme.applyTheme()
- Update VToggle: green on-state (systemPositiveStrong), correct off/disabled colors
- Migrate all pill VTabs call sites to VSegmentControl
