+++ clients/macos/vellum-assistant/Features/MainWindow/Panels/SharedFileViewerComponents.swift
@@ -89,10 +89,9 @@ struct FileContentView: View {
                 fileName: fileName
             ) {
                 if modes.count > 1 {
-                    VTabs(
+                    VSegmentControl(
                         items: modes.map { (label: viewModeLabel($0), tag: $0) },
                         selection: $viewMode,
-                        style: .pill
                     )
                     .fixedSize()
                 }
