+++ clients/macos/vellum-assistant/Features/Settings/ServiceModeCard.swift
@@ -50,13 +50,12 @@ struct ServiceModeCard<ManagedContent: View, YourOwnContent: View>: View {
                         .foregroundStyle(VColor.contentTertiary)
                 }
                 Spacer()
-                VTabs(
+                VSegmentControl(
                     items: [
                         (label: "Managed", tag: "managed"),
                         (label: "Your Own", tag: "your-own"),
                     ],
                     selection: $draftMode,
-                    style: .pill
                 )
                 .frame(width: 220)
             }
